+++ src/papilo/presolvers/SimplifyInequalities.hpp
@@ -334,7 +334,7 @@ SimplifyInequalities<REAL>::execute( const Problem<REAL>& problem,
       if( result == PresolveStatus::kUnchanged )
          return PresolveStatus::kUnchanged;
 
-      for( int i = 0; i < stored_reductions.size(); ++i )
+      for( int i = 0; i < (int) stored_reductions.size(); ++i )
       {
          Reductions<REAL> reds = stored_reductions[i];
          if( reds.size() > 0 )
